refactor(mcp): rename feedback module files to match the API stem
agent-feedback.ts -> feedback.ts (and the test file), plus the internal
description constant; "agent feedback" stays as the prose term. The ADR
filename keeps its original slug.

Generated-By: PostHog Desktop
Task-Id: 750e73c1-70eb-4697-a51a-840b207594b1
