+++ packages/mcp/src/extensions/posthog-mcp.ts
@@ -23,7 +23,7 @@ import {
   parseFeedbackReport,
   resolveCollectFeedbackOptions,
   SEND_FEEDBACK_TOOL_NAME,
-} from './agent-feedback'
+} from './feedback'
 import { analyticsOwnsParameter, stripOwnedAnalyticsArguments } from './analytics-parameters'
 import {
   addContextParameterToTools,
